Add lunar_pole_exploration_rover demo (#36).
This commit adds a lunar pole exploration rover demo modelled on NASA's VIPER
mission, including a simulated rover model, and world and ground models to
simulate the Mons Mouton.

The commit also adds Gazebo plugins to simulate power generation and
consumption in space robotics, which the simulation above uses. Plugins
include:

- A solar panel plugin to simulate power generation from the sun according to
  occlusion and angle between the sun and the panel.

- A radioisotope thermal generator plugin to simulate the constant power from a
  radioisotope thermal generator.

- A modified version of the linear battery plugin that is able to take as
  charge input the power outputs of the two previous plugins.

- A sensor power load system plugin to simulate the power drawn by sensors

Documentation, as well as a dockerfile to help build and try this demo, are
included.

Co-authored-by: stevedan <stevedan.o.omodolor@gmail.com>